+++ contracts/indexer/src/handlers/mod.rs
@@ -141,7 +141,7 @@ pub fn handle_log(conn: &mut PgConnection, log: Log, provider: &impl LogsProvide
 }
 
 #[cfg(test)]
-mod tests {
+mod test_utils {
     mod test_db;
     mod test_provider;
